Fix onion message tests by changing the rate limiter (#2132)
We want to be able to send several messages immediately after opening the connection which is not supported by Google's rate limiter that assumes that we had already reached the maximum rate when the rate limiter was created.
We create our own rate limiter that does what we need.

+++ eclair-core/src/main/scala/fr/acinq/eclair/NodeParams.scala
@@ -461,7 +461,7 @@ object NodeParams extends Logging {
         pingDisconnect = config.getBoolean("peer-connection.ping-disconnect"),
         maxRebroadcastDelay = FiniteDuration(config.getDuration("router.broadcast-interval").getSeconds, TimeUnit.SECONDS), // it makes sense to not delay rebroadcast by more than the rebroadcast period
         killIdleDelay = FiniteDuration(config.getDuration("onion-messages.kill-transient-connection-after").getSeconds, TimeUnit.SECONDS),
-        maxOnionMessagesPerSecond = config.getDouble("onion-messages.max-per-peer-per-second")
+        maxOnionMessagesPerSecond = config.getInt("onion-messages.max-per-peer-per-second")
       ),
       routerConf = RouterConf(
         channelExcludeDuration = FiniteDuration(config.getDuration("router.channel-exclude-duration").getSeconds, TimeUnit.SECONDS),
